+++ c-api/resvg.h
@@ -17,9 +17,9 @@
 #include <stdint.h>
 
 #define RESVG_MAJOR_VERSION 0
-#define RESVG_MINOR_VERSION 26
-#define RESVG_PATCH_VERSION 1
-#define RESVG_VERSION "0.26.1"
+#define RESVG_MINOR_VERSION 27
+#define RESVG_PATCH_VERSION 0
+#define RESVG_VERSION "0.27.0"
 
 /**
  * @brief List of possible errors.